+++ src/components/features/Posts/index.tsx
@@ -1,6 +1,7 @@
 'use client'
 import { getContentsApi } from '@/feature/cms/hooks/MicroCmsContents'
 import type { PostsType, WorksType } from '@/libs/cms/types/MicroCmsType'
+import { createLoadingPosts } from '@/libs/post/createPostDummy'
 import { type JSX, useEffect, useRef, useState } from 'react'
 import { LinkButton } from '../Button/LinkButton'
 import { Post } from '../Post'
@@ -9,6 +10,8 @@ import styles from './index.module.scss'
 export const Posts = () => {
   const [postElement, setPostElement] = useState<JSX.Element[]>([])
   const isFirstRender = useRef(true)
+  const CONTENTS_PER_PAGE = 5
+  const DUMMY_ELEMENT = createLoadingPosts(CONTENTS_PER_PAGE)
 
   useEffect(() => {
     if (!isFirstRender.current) return
@@ -48,7 +51,7 @@ export const Posts = () => {
 
   return (
     <div className={styles.container}>
-      {postElement}
+      {postElement.length === 0 ? DUMMY_ELEMENT : postElement}
       <LinkButton className={`${styles.button}`} text="もっと見る" link="/posts" />
     </div>
   )
