+++ src/libs/post/index.module.scss
@@ -1,0 +1,7 @@
+.dummy {
+  visibility: hidden;
+}
+
+.loader {
+  position: absolute;
+}
